Update tree-shaking.md
As a conclusion, I thought it would be a good idea to show case a high-level explanation of why it's called tree shaking. Once the user reads the guide and ends up with a high-level explanation I think everything will make much more sense, and the remainder of the word tree shaking will stick to the user's head.

If you think this is a good idea and believe there's a better place to showcase the high-level explanation, please go ahead and do it.